Parking space commands ready
parkingspace commands ready

+++ assignment_1/lib/models/parking_space.dart
@@ -1,7 +1,7 @@
 class Parkingspace {
   String id;
   String adress;
-  double pricePerHour;
+  int pricePerHour;
 
   Parkingspace(this.id, this.adress, this.pricePerHour);
 }
